Splitting of `ydb/library/yql/providers/generic/connector/tests` into different subfolders (#3172)
