+++ ydb/library/yql/providers/generic/connector/tests/datasource/clickhouse/docker-compose.yml
@@ -1,17 +1,8 @@
 version: '3.4'
 services:
-  postgresql:
-    image: postgres:15-bullseye@sha256:3411b9f2e5239cd7867f34fcf22fe964230f7d447a71d63c283e3593d3f84085
-    container_name: ${USER}_connector-integration-tests-postgresql
-    environment:
-      POSTGRES_DB: db
-      POSTGRES_USER: user
-      POSTGRES_PASSWORD: password
-    ports:
-      - 5432
   clickhouse:
     image: clickhouse/clickhouse-server:23-alpine@sha256:b078c1cd294632afa2aeba3530e7ba2e568513da23304354f455a25fab575c06
-    container_name: ${USER}_connector-integration-tests-clickhouse
+    container_name: fq-tests-ch-clickhouse
     environment:
       CLICKHOUSE_DB: db
       CLICKHOUSE_USER: user
@@ -21,9 +12,9 @@ services:
       - 9000
       - 8123
   fq-connector-go:
-    container_name: ${USER}_connector-integration-tests-fq-connector-go
+    container_name: fq-tests-ch-fq-connector-go
     image: ghcr.io/ydb-platform/fq-connector-go:v0.2.5@sha256:7f086ce3869b84a59fd76a10a9de8125c0d382915e956d34832105e03829a61b
     volumes:
-      - ./fq-connector-go/config.debug.yaml:/opt/ydb/cfg/fq-connector-go.yaml
+      - ../../fq-connector-go/:/opt/ydb/cfg/
     ports:
       - 2130
