+++ ydb/library/yql/providers/generic/connector/tests/datasource/ya.make
@@ -1,0 +1,4 @@
+RECURSE_FOR_TESTS(
+    clickhouse
+    postgresql
+)
