+++ ydb/library/yql/providers/generic/connector/tests/fq-connector-go/fq-connector-go.yaml
