+++ ydb/library/yql/providers/generic/connector/tests/test_cases/ya.make
@@ -4,15 +4,9 @@ STYLE_PYTHON()
 
 PY_SRCS(
     base.py
-    collection.py
-    join.py
-    select_datetime.py
     select_missing_database.py
     select_missing_table.py
-    select_positive_clickhouse.py
     select_positive_common.py
-    select_positive_postgresql.py
-    select_positive_postgresql_schema.py
 )
 
 PEERDIR(
