+++ ydb/library/yql/providers/generic/connector/tests/utils/types/ya.make
@@ -1,0 +1,10 @@
+PY3_LIBRARY()
+
+STYLE_PYTHON()
+
+PY_SRCS(
+    clickhouse.py
+    postgresql.py
+)
+
+END()
