+++ ydb/library/yql/providers/generic/provider/yql_generic_provider.cpp
@@ -21,7 +21,6 @@ namespace NYql {
             Y_UNUSED(randomProvider);
             Y_UNUSED(progressWriter);
             Y_UNUSED(operationOptions);
-            Y_ENSURE(credentialsFactory);
 
             auto state = MakeIntrusive<TGenericState>(
                 typeCtx.Get(),
